backend/requirements.txt
asgiref==3.7.2
build==1.0.3
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
jwcrypto==1.5.0
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pip-tools==7.3.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==68.2.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wheel==0.41.2
wrapt==1.15.0
